+++ src/base_dumper/common/dump_format.py
@@ -4,6 +4,5 @@ from enum import Enum
 class DumpFormat(Enum):
     """Enum for dump type format."""
 
-    RAW = 0
-    CSV = 1
-    S3 = 2
+    binary = 0
+    csv = 1