+++ src/base_dumper/common/stream.py
@@ -1,0 +1,6 @@
+STREAM_TYPE = {
+    "clickhouse": "native",
+    "greenplum": "pgcopy",
+    "postgres": "pgcopy",
+    "sqlserver": "bcp",
+}